--- a/experimental/segmenter/src/complex.rs
+++ b/experimental/segmenter/src/complex.rs
@@ -162,7 +162,7 @@ pub fn complex_language_segment_utf16(
             if let Some(lstm) = lstm {
                 if let Some(model) = lstm.best(*first_ch as u32) {
                     if let Ok(segmenter) = LstmSegmenter::try_new_unstable(model, grapheme) {
-                        let breaks = segmenter.segment_utf16(&str_per_lang);
+                        let breaks = segmenter.segment_utf16(str_per_lang);
                         result.extend(breaks.map(|n| offset + n));
                         offset += str_per_lang.len();
                         result.push(offset);
@@ -177,7 +177,7 @@ pub fn complex_language_segment_utf16(
                         if let Ok(segmenter) =
                             DictionarySegmenter::try_new_unstable(payload, grapheme)
                         {
-                            let breaks = segmenter.segment_utf16(&str_per_lang);
+                            let breaks = segmenter.segment_utf16(str_per_lang);
                             result.extend(breaks.map(|n| offset + n));
                             offset += str_per_lang.len();
                             continue;
@@ -210,7 +210,7 @@ pub fn complex_language_segment_str(
             if let Some(lstm) = lstm {
                 if let Some(model) = lstm.best(first_ch as u32) {
                     if let Ok(segmenter) = LstmSegmenter::try_new_unstable(model, grapheme) {
-                        let breaks = segmenter.segment_str(&str_per_lang);
+                        let breaks = segmenter.segment_str(str_per_lang);
                         result.extend(breaks.map(|n| offset + n));
                         offset += str_per_lang.chars().map(|c| c.len_utf8()).sum::<usize>();
                         result.push(offset);
@@ -225,7 +225,7 @@ pub fn complex_language_segment_str(
                         if let Ok(segmenter) =
                             DictionarySegmenter::try_new_unstable(payload, grapheme)
                         {
-                            let breaks = segmenter.segment_str(&str_per_lang);
+                            let breaks = segmenter.segment_str(str_per_lang);
                             result.extend(breaks.map(|n| offset + n));
                             offset += str_per_lang.chars().map(|c| c.len_utf8()).sum::<usize>();
                             continue;