Reducing string allocations. Mostly noticeable in dictionary mode.

```
Line Break/Latin1/En    time:   [5.6400 µs 5.6509 µs 5.6627 µs]                                  
                        change: [-0.7285% -0.4760% -0.2138%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
Line Break/Latin1/En CSS                                                                             
                        time:   [4.8874 µs 4.8957 µs 4.9049 µs]
                        change: [+3.5165% +4.0746% +4.5868%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

Line Break/UTF8/En      time:   [12.162 µs 12.177 µs 12.195 µs]                                
                        change: [-6.1528% -5.9502% -5.7570%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe
Line Break/UTF8/En CSS  time:   [11.701 µs 11.722 µs 11.745 µs]                                    
                        change: [-1.7091% -1.2712% -0.8771%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  9 (9.00%) high mild
  5 (5.00%) high severe
Line Break/UTF8/Th/auto time:   [494.21 µs 494.79 µs 495.47 µs]                                    
                        change: [+0.9365% +1.1126% +1.2930%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
Line Break/UTF8/Th/lstm time:   [494.60 µs 495.17 µs 495.82 µs]                                    
                        change: [+1.0661% +1.2626% +1.4677%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 14 outliers among 100 measurements (14.00%)
  10 (10.00%) high mild
  4 (4.00%) high severe
Line Break/UTF8/Th/dictionary                                                                             
                        time:   [4.6666 µs 4.6723 µs 4.6790 µs]
                        change: [-13.748% -13.580% -13.405%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

Line Break/UTF16/En     time:   [11.965 µs 12.018 µs 12.100 µs]                                 
                        change: [-11.108% -10.548% -10.033%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
Line Break/UTF16/En CSS time:   [11.434 µs 11.465 µs 11.498 µs]                                     
                        change: [-10.562% -10.284% -9.9902%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
Line Break/UTF16/Th/auto                                                                            
                        time:   [492.07 µs 492.68 µs 493.29 µs]
                        change: [-0.9509% -0.0908% +0.6572%] (p = 0.85 > 0.05)
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) high mild
  8 (8.00%) high severe
Line Break/UTF16/Th/lstm                                                                            
                        time:   [493.69 µs 494.34 µs 494.98 µs]
                        change: [+1.1125% +1.2662% +1.4013%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
Line Break/UTF16/Th/dictionary                                                                             
                        time:   [3.4984 µs 3.5043 µs 3.5113 µs]
                        change: [-11.808% -11.260% -10.758%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe
```